services/tts-openvoice/requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
babel==2.18.0
boto3==1.42.53
botocore==1.42.53
brotli==1.2.0
cached_path==1.8.9
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cn2an==0.5.23
colorama==0.4.6
cryptography==46.0.5
csvw==3.7.0
ctranslate2==4.7.1
Cython==3.2.4
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.3.1
Distance==0.1.3
dlinfo==2.0.0
docopt==0.6.2
dtw-python==1.7.4
eng_to_ipa==0.0.2
fastapi==0.129.0
faster-whisper==1.2.1
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.12.19
fsspec==2025.12.0
fugashi==1.5.2
g2p-en==2.1.0
g2pK==0.9.4
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gradio==6.6.0
gradio_client==2.1.0
groovy==0.1.2
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
inflect==7.5.0
isodate==0.7.2
jaconv==0.5.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jpype1==1.6.0
jsonlines==1.2.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
konlpy==0.6.0
langid==1.1.6
language-tags==1.2.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mecab-python3==1.0.12
melotts @ git+https://github.com/myshell-ai/MeloTTS.git@209145371cff8fc3bd60d7be902ea69cbdb7965a
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
# OpenVoice se instala como editable desde el repo local:
#   pip install -e ./repo/OpenVoice --no-deps
# (evitar --no-deps omite versiones obsoletas en setup.py de OpenVoice)
-e ./repo/OpenVoice
networkx==3.6.1
nltk==3.9.2
num2words==0.5.14
numba==0.64.0
numpy==2.4.2
onnxruntime==1.24.2
openai-whisper==20250625
orjson==3.11.7
packaging==26.0
pandas==3.0.1
phonemizer==3.3.0
pillow==12.1.1
pip==26.0.1
plac==1.4.5
platformdirs==4.9.2
pooch==1.9.0
proces==0.1.7
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pykakasi==2.3.0
pyparsing==3.3.2
pypinyin==0.55.0
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
resampy==0.4.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
segments==2.3.0
semantic-version==2.10.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
starlette==0.52.1
sympy==1.14.0
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
tqdm==4.67.3
transformers==5.2.0
txtsplit==1.0.0
typeguard==4.5.1
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
Unidecode==1.4.0
unidic==1.1.0
unidic-lite==1.0.8
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
wasabi==0.10.1
wavmark==0.0.3
whisper-timestamped==1.15.9
wrapt==2.1.1
